pemisahan controller dengan routes dari user
